Added new library and reorganized tests
