+++ app/platformio.ini
@@ -13,12 +13,12 @@ default_envs = teensy41
 
 [env]
 lib_ldf_mode = chain+
-lib_ignore =
-    ; not fully implemented
-    Recorder
+lib_ignore = 
+	Recorder
 
 [env:native]
 platform = native
+lib_deps = https://github.com/FrankBoesing/Teensy-WavePlayer.git
 
 [env:teensy41]
 platform = teensy
@@ -27,3 +27,4 @@ framework = arduino
 upload_protocol = teensy-cli
 upload_speed = 9600
 build_flags = -D USB_MIDI_AUDIO_SERIAL
+lib_deps = https://github.com/FrankBoesing/Teensy-WavePlayer.git
